second commit, split up components
