+++ 2020/join/gatsby-new-join/src/components/intro.js
@@ -1,0 +1,13 @@
+import React from "react"
+
+const Intro = () => (
+  <section className="intro">
+    <div className="intro-text">
+      <h1>The Daily Pennsylvanian, Inc.</h1>
+      <h3>Your campus compass. Join us and find your direction.</h3>
+      <a href="#apply"><div className="join-button">Join Now</div></a>
+    </div>
+  </section>
+)
+
+export default Intro
