+++ 2020/join/gatsby-new-join/src/components/join.js
@@ -1,0 +1,17 @@
+import React from "react"
+
+const Join = () => (
+  <section className="join" id="apply">
+    <h2>Join The Daily Pennsylvanian, Inc.</h2>
+    <div className="section form">
+      <p className="overall">
+        Fill out the form below to apply!
+      </p>
+      <iframe className="join-form" src="https://docs.google.com/forms/d/e/1FAIpQLSfXatYm_DWJOMUTtFgs45uQar2g1ycqXuX-YeUGb5q0McueMg/viewform" width="100%" height="600" frameBorder="0" marginHeight="0" marginWidth="0">
+        Loading...
+      </iframe>
+    </div>
+  </section>
+)
+
+export default Join
