+++ 2020/join/gatsby-new-join/src/components/street.js
@@ -1,0 +1,36 @@
+import React from "react"
+
+const Street = () => (
+  <section className="street">
+    <h2>34th Street Magazine</h2>
+    <div className="section">
+      <p>
+        34th Street Magazine is your authority on Penn student life. We are a student magazine that empowers student voices, sheds light on student issues, and provides timely commentary on the arts and campus culture. We provide endless opportunities for a creative outlet: you can write cover stories, help envision marketing campaigns, design and lay out issues, attend press movie screenings, review top-rated restaurants, and even pitch your own content ideas.
+      </p>
+      <p>
+        As the hub for Penn voices and narrative, Street staffers cover arts, culture, and feature content. We are looking for writers, photographers, videographers, artists, and social media whizzes to create, package, and promote Street to the Penn audience. Writers can work for our Features, Style, Arts, Film &amp; TV, Music, or Ego sections. Multimedia staffers work with either the Photo or Video departments. Illustrators and graphic designers create visuals, including our weekly print product, as part of the Design department. Audience Engagement associates manage our social media and our newsletter, The Toast.
+      </p>
+      <p>
+        Street provides writers the opportunity to be creative in their pursuit of journalism, teaches Audience Engagement staff how to build and promote a brand online, and gives photographers, videographers, and Design associates experience in creating engaging editorial visuals. Join Street to be part of a tight–knit community and learn from some of the most interesting and dedicated people at Penn—we promise you won’t regret it.
+      </p>
+      <h5>Highlights</h5>
+      <a href="https://www.34st.com/article/2019/11/hong-kong-protests-protestors-penn-ivy-league-extradition-bill-students-government-china-beijing-revolution-police-activists-politics" target="_blank">
+        ‘Revolution of Our Times’: Penn Students from Hong Kong Reflect on the Protests →
+      </a>
+      <a href="https://www.34st.com/article/2019/11/climate-change-global-warming-college-upenn-fossil-free-penn" target="_blank">
+        The 2019 Climate Change Issue →
+      </a>
+      <a href="https://www.34st.com/article/2019/12/best-movies-decade-2010s-moonlight-carol-mad-max-get-out-grand-budapest-hotel-inside-llewyn-davis-lady-bird" target="_blank">
+        Street’s Favorite Movies of the Decade →
+      </a>
+      <a href="https://www.34st.com/article/2019/10/therapy-mental-health-emotions-feelings-anxiety-trauma" target="_blank">
+        Why I’m Going Back to Therapy →
+      </a>
+      <a href="https://www.34st.com/article/2019/12/dizengoff-frozen-labneh-solomonov-zahav-philadelphia-food-foodie-restaurants-israeli-cuisine-middle-eastern-kfar-dizengoff-goldie-university-city" target="_blank">
+        Dizengoff Just Opened on Campus—and Yes, We Tried the Frozen Labneh →
+      </a>
+    </div>
+  </section>
+)
+
+export default Street
