+++ 2020/join/gatsby-new-join/src/components/utb.js
@@ -1,0 +1,25 @@
+import React from "react"
+
+const UTB = () => (
+  <section className="utb">
+    <h2>Under the Button</h2>
+    <div className="section">
+      <p>
+        Under the Button is Penn’s go-to source for daily humor and satire.
+      </p>
+      <p>
+        Our writing staff works around the clock to report on the stories that other media organizations wouldn’t dare touch due to editorial standards. Our video and media staff work to capture the real Penn experience through the lens of a camera we stole from the DP. 
+      </p>
+      <p>
+        We are looking for talented writers, videographers, actors, graphic designers, coders, social media specialists, our dignity, ten gay rats, and your best brisket recipe. If you’re an all-around funny person looking for a way to subject the rest of campus to your sense of humor, we’re the publication for you.
+      </p>
+      <h5>Highlights</h5>
+      <a href="https://www.underthebutton.com/article/2019/07/wax-trump-shame-penn" target="_blank">Wax and Trump in Heated Competition to Bring Most Shame to Penn →</a>
+      <a href="https://www.underthebutton.com/article/2019/09/if-i-had-a-girlfriend-photo-essay" target="_blank">Photo Essay: These Are All the Places on Campus I Would Kiss My Girlfriend If I Had One →</a>
+      <a href="https://www.underthebutton.com/article/2018/12/choose-your-own-adventure-jeff-writing-sem-intro" target="_blank">Choose Your Own Adventure: Will You Fuck Jeff From Writing Sem? →</a>
+      <a href="https://youtu.be/Or5lGThQTJU" target="_blank"> Pool Party Man on the Street →</a>
+    </div>
+  </section>
+)
+
+export default UTB
